Allow playing hot or recent tracks instead of incoming tracks
Adds a separate track type option which stores key names ("incoming",
"hot" or "recent") in localStorage.type.
The type is then mapped to a URL (including parameters, terminated with
a ?) by trackTypeMapping object in background.js.
Because track feeds don't have the same layout as the
affiliated/incoming tracks, it checks if the response is an array or
not of whether to use the collection/origin parameters.